// project/src/models/enums/ItemEventActions.ts
export enum ItemEventActions
    {
    MOVE = "Move",
    REMOVE = "Remove",
    SPLIT = "Split",
    MERGE = "Merge",
    TRANSFER = "Transfer",
    SWAP = "Swap",
    FOLD = "Fold",
    TOGGLE = "Toggle",
    TAG = "Tag",
    BIND = "Bind",
    EXAMINE = "Examine",
    READ_ENCYCLOPEDIA = "ReadEncyclopedia",
    APPLY_INVENTORY_CHANGES = "ApplyInventoryChanges",
    CREATE_MAP_MARKER = "CreateMapMarker",
    DELETE_MAP_MARKER = "DeleteMapMarker",
    EDIT_MAP_MARKER = "EditMapMarker",
    OPEN_RANDOM_LOOT_CONTAINER = "OpenRandomLootContainer",
    HIDEOUT_QTE_EVENT = "HideoutQuickTimeEvent"
}